style: added vite bundler, and code clean up
