+++ public/images/color-palette.png
