+++ src/css/style.css
@@ -1,3 +1,5 @@
+@import 'tailwindcss';
+
 * {
   margin: 0;
   padding: 0;
