+++ src/main.js
@@ -1,0 +1,4 @@
+import "./css/styles.css";
+import { basicSetUP } from "./javascripts/script";
+
+basicSetUP();
